+++ src/main/java/io/github/asw/i3a/agentsWebClient/Service.java
@@ -1,0 +1,15 @@
+package io.github.asw.i3a.agentsWebClient;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * Hello world!
+ *
+ */
+@SpringBootApplication
+public class Service {
+    public static void main(String[] args) {
+	SpringApplication.run(Service.class, args);
+    }
+}
