+++ src/main/java/io/github/asw/i3a/agentsWebClient/types/Operator.java
@@ -1,0 +1,22 @@
+/*
+ * This source file is part of the Agents_i3a open source project.
+ *
+ * Copyright (c) 2017 Agents_i3a project authors.
+ * Licensed under MIT License.
+ *
+ * See /LICENSE for license information.
+ * 
+ * This class is based on the AlbUtil project.
+ * 
+ */
+package io.github.asw.i3a.agentsWebClient.types;
+
+
+import lombok.Data;
+
+@Data
+public class Operator {
+	private String operatorId;
+	private String name;
+	private String email;
+}
