--- src/main/java/org/uniovi/i3a/incimanager/Service.java
@@ -1,15 +1,0 @@
-package org.uniovi.i3a.incimanager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * Hello world!
- *
- */
-@SpringBootApplication
-public class Service {
-    public static void main(String[] args) {
-	SpringApplication.run(Service.class, args);
-    }
-}
