--- src/main/java/org/uniovi/i3a/incimanager/kafka/KafkaProduceFactory.java
@@ -1,58 +1,0 @@
-package org.uniovi.i3a.incimanager.kafka;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.apache.kafka.clients.CommonClientConfigs;
-import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.core.DefaultKafkaProducerFactory;
-import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.core.ProducerFactory;
-
-@Configuration
-public class KafkaProduceFactory {
-
-	@Value("${spring.kafka.producer.bootstrap-servers}")
-	String server;
-	@Value("${karafka.username}")
-	String username;
-	@Value("${karafka.password}")
-	String password;
-	@Value("${kafka.protocol}")
-	String protocol;
-
-	@Bean
-	public ProducerFactory<String, String> producerFactory() {
-		return new DefaultKafkaProducerFactory<>(producerConfigs());
-	}
-
-	@Bean
-	public Map<String, Object> producerConfigs() {
-		String jaasTemplate = "org.apache.kafka.common.security.scram.ScramLoginModule required username=\"%s\" password=\"%s\";";
-		String jaasCfg = String.format(jaasTemplate, username, password);
-
-		Map<String, Object> props = new HashMap<>();
-		props.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, server);
-		props.put("group.id", username + "-consumer");
-		props.put("enable.auto.commit", "true");
-		props.put("auto.commit.interval.ms", "1000");
-		props.put("auto.offset.reset", "earliest");
-		props.put("session.timeout.ms", "30000");
-		props.put(CommonClientConfigs.SECURITY_PROTOCOL_CONFIG, protocol);
-		props.put("sasl.mechanism", "SCRAM-SHA-256");
-		props.put("sasl.jaas.config", jaasCfg);
-		props.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
-		props.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class.getName());
-		return props;
-	}
-
-	@Bean
-	public KafkaTemplate<String, String> kafkaTemplate() {
-		return new KafkaTemplate<String, String>(producerFactory());
-	}
-
-}
