--- src/main/java/org/uniovi/i3a/incimanager/kafka/package-info.java
@@ -1,1 +1,0 @@
-package org.uniovi.i3a.incimanager.kafka;
