--- src/main/java/org/uniovi/i3a/incimanager/types/Comment.java
@@ -1,12 +1,0 @@
-package org.uniovi.i3a.incimanager.types;
-
-import java.util.Date;
-
-import lombok.Data;
-
-@Data
-public class Comment {
-    private Date date = new Date();
-    private String operatorId = "";
-    private String comment = "";
-}
