--- src/main/java/org/uniovi/i3a/incimanager/web/IncidentInfo.java
@@ -1,18 +1,0 @@
-package org.uniovi.i3a.incimanager.web;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class IncidentInfo {
-	private String title;
-	private String description;
-	private String location;
-	private String operatorId;
-	private String expirationDate;
-	private String tags;
-	private String multimedia;
-	private String properties;
-	private String status;
-}
