+++ src/main/resources/application-test.properties
@@ -1,5 +1,6 @@
 server.port=8090
-agents.url=http://localhost:8080/user
+agents.url=http://asw-i3a-zuul-eu-west-1.guill.io/agents_service
+agents.url=http://asw-i3a-zuul-eu-west-1.guill.io/incidents_service
 
 incidence.defaults.state=OPEN
