+++ src/main/resources/application.properties
@@ -4,7 +4,9 @@ eureka.name=agents_web_client
 eureka.client.serviceUrl.defaultZone=${DEFAULT_EUREKA:http://asw-i3a-eureka-eu-west-1.guill.io/eureka/}
 eureka.instance.prefer-ip-address=true
 
-agents.url=https://i3a-2.herokuapp.com/user
+agents.url=http://asw-i3a-zuul-eu-west-1.guill.io/agents_service
+incidents.url=http://asw-i3a-zuul-eu-west-1.guill.io/incidents_service
+operators.url=http://asw-i3a-zuul-eu-west-1.guill.io/operators_service
 
 spring.application.name=agents_web_client
