+++ src/main/resources/static/style.css
@@ -5,6 +5,7 @@ html {
 }
 
 .aws-page-header {
+  padding-bottom: 5px;
   border-bottom: 1px solid #e0e0e0;
 }
