--- src/test/java/org/uniovi/i3a/incimanager/web/WebControllerTest.java
@@ -1,61 +1,0 @@
-/*
- * This source file is part of the rest-service open source project.
- *
- * Copyright (c) 2018 willy and the rest-service project authors.
- * Licensed under GNU General Public License v3.0.
- *
- * See /LICENSE for license information.
- * 
- */
-package org.uniovi.i3a.incimanager.web;
-
-import static org.junit.Assert.*;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.mock.web.MockHttpSession;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.uniovi.i3a.incimanager.Service;
-
-import TestKit.IntegrationTest;
-
-/**
- * Instance of RestControllerTest.java
- * 
- * @author
- * @version
- */
-@SpringBootTest(classes = { Service.class })
-@RunWith(SpringJUnit4ClassRunner.class)
-@Category(IntegrationTest.class)
-@ActiveProfiles("test")
-public class WebControllerTest {
-
-    @Autowired
-    private WebApplicationContext context;
-    @SuppressWarnings("unused")
-    private MockHttpSession session;
-    @SuppressWarnings("unused")
-    private MockMvc mockMvc;
-
-    @Before
-    public void setUp() throws Exception {
-
-	this.mockMvc = MockMvcBuilders.webAppContextSetup(this.context).build();
-
-	session = new MockHttpSession();
-    }
-    
-    @Test
-    public void pass() {
-	assertEquals(true, true);
-    }
-}
